Move and re-enable composition tests
Moved many of the composition tests to be unit tests
instead of integration tests and removed the dependency
on android views.

Test: ./gradles :compose:r:r:tDUT
Bug: b/179279455
