+++ compose/runtime/runtime/src/test/kotlin/androidx/compose/runtime/CompositionTests.kt
@@ -35,6 +35,7 @@ import androidx.compose.runtime.mock.SelectContact
 import androidx.compose.runtime.mock.compositionTest
 import androidx.compose.runtime.mock.skip
 import androidx.compose.runtime.mock.Text
+import androidx.compose.runtime.mock.View
 import androidx.compose.runtime.mock.expectChanges
 import androidx.compose.runtime.mock.expectNoChanges
 import androidx.compose.runtime.mock.validate
@@ -2804,7 +2805,7 @@ class CompositionTests {
 
 @OptIn(InternalComposeApi::class, ExperimentalComposeApi::class)
 @Composable
-private fun TestSubcomposition(
+internal fun TestSubcomposition(
     content: @Composable () -> Unit
 ) {
     val parentRef = rememberCompositionContext()
@@ -2817,6 +2818,32 @@ private fun TestSubcomposition(
         subcomposition.applyChanges()
         onDispose {
             subcomposition.dispose()
+        }
+    }
+}
+
+class Ref<T : Any> {
+    lateinit var value: T
+}
+
+@Composable fun NarrowInvalidateForReference(ref: Ref<CompositionContext>) {
+    ref.value = rememberCompositionContext()
+}
+
+@Composable
+fun testDeferredSubcomposition(block: @Composable () -> Unit): () -> Unit {
+    val container = remember { View() }
+    val ref = Ref<CompositionContext>()
+    NarrowInvalidateForReference(ref = ref)
+    return {
+        @OptIn(ExperimentalComposeApi::class)
+        Composition(
+            ViewApplier(container),
+            ref.value
+        ).apply {
+            setContent {
+                block()
+            }
         }
     }
 }
